fixed big when deleting a job with lots of tiles in the temp tile bucket
When cancelling a job that has backed up and has lots of tiles in the temp tile bucket, it can take a few minutes to delete. Added small tweaks to stop polling the job during the delete and disable the cancel button during the delete. On success user is redirected to the listing