Prevent 500 error when there's a failed downsample.
Previously, if a downsample's step function failed, the DB would have
that channel marked as downsample in progress and hold the ARN of that
step function.  When a user tried to do a new downsample, an exception
would be thrown if that failed step function ARN no longer exists.

Now, if that exception is thrown, the channel's downsample status is
marked as failed and the ARN of that non-existent step function is
removed from the DB.