Fix error creating channel via form as non-admin.
Removed default value for the bucket name from the form since only
admins allowed to specify the bucket name and we allow `null` in the DB
for backwards compatibility.

Make default value of bucket and cv_path None.  Since these fields may
be null in the DB, use None as the default value for them.